verifying router, image process

+++ routers/posts.py
@@ -8,6 +8,7 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import *
+from auth import CurrentUser
 router = APIRouter()
 
 @router.post(
@@ -15,21 +16,15 @@ router = APIRouter()
     response_model=ResponsePost,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: CreatePost, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User).where(models.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(
+    post: CreatePost, 
+    current_user:CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id = current_user.id
     )
     db.add(new_post)
     await db.commit()
@@ -38,7 +33,7 @@ async def create_post(post: CreatePost, db: Annotated[AsyncSession, Depends(get_
 
 #-------------------------- DELETE POST------------------
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
@@ -46,6 +41,11 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
         )
+    if post.user_id != current_user.id:
+                raise HTTPException(
+                    status_code=status.HTTP_403_FORBIDDEN,
+                    detail="Not Authorized to update this post",
+                )
 
     await db.delete(post)
     await db.commit()
@@ -84,6 +84,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 async def update_post_partial(
     post_id: int,
     post_data: UpdatePost,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -93,7 +94,12 @@ async def update_post_partial(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
         )
-
+    if post.user_id != current_user.id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Not Authorized to update this post",
+            )
+            
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
         setattr(post, field, value)
@@ -107,6 +113,7 @@ async def update_post_partial(
 async def update_post_full(
     post_id: int,
     post_data: CreatePost,
+    current_user:CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -126,10 +133,15 @@ async def update_post_full(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail="User not found",
             )
-
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to update this post",
+        )
+        
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    post.user_id = current_user.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
